rename naive -> tinyblas
