+++ llamafile/cuda.c
@@ -34,8 +34,8 @@
 __static_yoink("llama.cpp/ggml.h");
 __static_yoink("llamafile/compcap.cu");
 __static_yoink("llamafile/llamafile.h");
-__static_yoink("llama.cpp/naive-gemm.h");
-__static_yoink("llama.cpp/naive-gemm.cu");
+__static_yoink("llama.cpp/tinyblas.h");
+__static_yoink("llama.cpp/tinyblas.cu");
 __static_yoink("llama.cpp/ggml-cuda.h");
 __static_yoink("llama.cpp/ggml-cuda.cu");
 
@@ -62,8 +62,8 @@ static const struct Source {
     {"/zip/llamafile/compcap.cu", "compcap.cu"},
     {"/zip/llamafile/llamafile.h", "llamafile.h"},
     {"/zip/llama.cpp/ggml-cuda.h", "ggml-cuda.h"},
-    {"/zip/llama.cpp/naive-gemm.h", "naive-gemm.h"},
-    {"/zip/llama.cpp/naive-gemm.cu", "naive-gemm.cu"},
+    {"/zip/llama.cpp/tinyblas.h", "tinyblas.h"},
+    {"/zip/llama.cpp/tinyblas.cu", "tinyblas.cu"},
     {"/zip/llama.cpp/ggml-cuda.cu", "ggml-cuda.cu"}, // must come last
 };
